🚚 Dataloader References (For Relational Data)
One of the biggest downside of GraphQL Resolvers its that it didn't
actually solve the overfetching issues. It mostly moved it inside the
server for Relational Data instead of over the network and the client.
This is definitely still better than REST but create what's called the
n+1 problem. A way to solve this is using dataloader to cache and
resolve multiple fetches into one single function or instruction.

This is my implementation of dataloader for gqlgen, but this is not
somthing that should be use right of the bat, rather more of a
reference.

Reasoning includes Dataloaden limitation with no generics. Context
specific middleware, completely different relational data and their
databases.

- Added Dataloaden to handle and resolve multiple queries of the same
type
- Implemented Dataloaden Middleware into Gin
- Updated data entities and databases connection to relational one
- Added UserService for handle queries for User
- Modified ThoughtService to take account for relational data and
  UserService
- Modified Fx dependecies structure to take account for new Providers
  and different dependecies